test(snapshot): share chunk descriptor fixture across builder and planner tests
createChunkBytes and createDescriptor were duplicated between
snapshot-chunk-builder.test.ts and chunked-fold-planner.test.ts, differing
only in whether the chunk incarnation was overridable. Both files landed in
this PR, so this extracts them before a third copy calcifies.

The fixture is a factory bound to one collection, so all ~49 existing call
sites are unchanged; collection, collectionPrefix, and incarnation are
arguments rather than module constants because each test file defines its
own.

snapshot-view.test.ts keeps its own copy deliberately. That block is
embedded in a seeder that also PUTs bytes into MemoryStorage and assembles
a SnapshotManifest, so extracting it would couple a view-layer fixture
seeder to a descriptor factory.
